# Generic event conversion of pointers, arrays, slices