Changes
- Dereference any pointers in the event.
- Apply generic event conversion to arrays and slices in the event.
- Drop nil values from MapStr. This does not apply to arrays, structs, or map types converted via JSON marshal/unmarshal.
- Drop keys containing values of unknown types (previously it only logged a warning).
- Provide a detailed warning when the generic event conversion is unsuccessful. A message is logged contained the map key that failed, the cause, and the event (for context).
- Due to the enhancement to normalize types in arrays, a change was required in Packetbeat's memcache proto.
- Changed a test case in Filebeat to not expect a null value.