@tsg 2016-08-26
@andrewkroh Consider also adding a benchmark with a few common conversions, so we have a performance baseline for future changes. I am a bit worried about the pointer following calling reflect for all values, but hopefully the cost of it is small enough.