@andrewkroh 2016-08-31
> @andrewkroh Consider also adding a benchmark with a few common conversions, so we have a performance baseline for future changes. I am a bit worried about the copying and the pointer following calling reflect for all values, but hopefully the cost of it is small enough.

I changed up the code to only attempt to deference the pointer after testing for other types first. And add some benchmarks for comparisons.

```
# Before
BenchmarkConvertToGenericEventNetString-4           	 3000000	       494 ns/op	     376 B/op	       4 allocs/op
BenchmarkConvertToGenericEventMapStringString-4     	  500000	      4104 ns/op	    1784 B/op	      23 allocs/op
BenchmarkConvertToGenericEventMapStr-4              	  300000	      4273 ns/op	    1808 B/op	      25 allocs/op
BenchmarkConvertToGenericEventStringSlice-4         	 3000000	       496 ns/op	     400 B/op	       4 allocs/op
BenchmarkConvertToGenericEventCustomStringSlice-4   	 3000000	       497 ns/op	     400 B/op	       4 allocs/op

# After
BenchmarkConvertToGenericEventNetString-4           	 2000000	       904 ns/op	     736 B/op	       8 allocs/op
BenchmarkConvertToGenericEventMapStringString-4     	  300000	      4167 ns/op	    2120 B/op	      25 allocs/op
BenchmarkConvertToGenericEventMapStr-4              	 1000000	      1443 ns/op	    1360 B/op	       9 allocs/op
BenchmarkConvertToGenericEventStringSlice-4         	 2000000	       790 ns/op	     736 B/op	       6 allocs/op
BenchmarkConvertToGenericEventCustomStringSlice-4   	 1000000	      1504 ns/op	     882 B/op	      15 allocs/op
BenchmarkConvertToGenericEventStringPointer-4       	 2000000	       845 ns/op	     688 B/op	       5 allocs/op

# Modifying the input map (not returning a copy)
# Things will get faster if we can track down
# what causes the race condition.
BenchmarkConvertToGenericEventNetString-4           	 2000000	       676 ns/op	     400 B/op	       6 allocs/op
BenchmarkConvertToGenericEventMapStringString-4     	  500000	      3857 ns/op	    1784 B/op	      23 allocs/op
BenchmarkConvertToGenericEventMapStr-4              	 1000000	      1013 ns/op	     688 B/op	       5 allocs/op
BenchmarkConvertToGenericEventStringSlice-4         	 3000000	       561 ns/op	     400 B/op	       4 allocs/op
BenchmarkConvertToGenericEventCustomStringSlice-4   	 1000000	      1315 ns/op	     546 B/op	      13 allocs/op
BenchmarkConvertToGenericEventStringPointer-4       	 2000000	       635 ns/op	     352 B/op	       3 allocs/op
```