@andrewkroh 2016-08-31
This is the race condition I'm seeing if I modify `ConvertToGenericEvent()` to modify the input parameter (vs returning a copy of the map). http://pastebin.com/raw/KCL0mTkH