@tsg 2016-09-01
@andrewkroh Thanks for doing the tests. I'd say that the performance numbers are good enough that we don't really have to worry about, most likely the bottleneck is going to be somewhere else. And if the profiling shows that it's here, people can do the conversion in the Beat before passing it to us.