@andrewkroh 2016-09-02
Based on discussion with urso about the stack trace from the race detector, `ConvertToGenericEvent` will continue to return a copy. So this PR is read to be squashed and merged (assuming there are no new comments).