@tsg 2016-09-02
Merged. Great work @andrewkroh, I'm really glad we'll have this in beta1.